separate auth-protected pages from login page
